# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad View Rewards API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/gift-codes": {
      "post": {
        "tags": [
          "admin-gift-codes"
        ],
        "summary": "Upload Gift Codes",
        "operationId": "upload_gift_codes_api_v1_admin_gift_codes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "amazon"
              ],
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Gift codes uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GiftCodeRead"
                  },
                  "title": "Response Upload Gift Codes Api V1 Admin Gift Codes Post"
                },
                "examples": {
                  "uploaded": {
                    "value": [
                      {
                        "id": "acecf050-e315-4db0-99ca-fdc7af77ec19",
                        "code": "CODE-001",
                        "provider": "amazon",
                        "redeemed": false,
                        "created_at": "2026-10-05T14:00:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "json_with_provider": {
                  "value": {
                    "provider": "amazon",
                    "codes": [
                      "CODE-001",
                      "CODE-002"
                    ]
                  }
                },
                "json_with_query_provider": {
                  "value": [
                    "CODE-001",
                    "CODE-002"
                  ]
                }
              }
            },
            "text/csv": {
              "examples": {
                "csv": {
                  "value": "code\nCODE-001\nCODE-002"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin-gift-codes"
        ],
        "summary": "List Gift Codes",
        "operationId": "list_gift_codes_api_v1_admin_gift_codes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "redeemed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                false
              ],
              "title": "Redeemed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gift code list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GiftCodeRead"
                  },
                  "title": "Response List Gift Codes Api V1 Admin Gift Codes Get"
                },
                "examples": {
                  "codes": {
                    "value": [
                      {
                        "id": "acecf050-e315-4db0-99ca-fdc7af77ec19",
                        "code": "CODE-001",
                        "provider": "amazon",
                        "redeemed": false,
                        "created_at": "2026-10-05T14:00:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/redeem-code": {
      "post": {
        "tags": [
          "admin-gift-codes"
        ],
        "summary": "Assign Gift Code",
        "operationId": "assign_gift_code_api_v1_admin_redeem_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GiftCodeRedeemRequest",
                "examples": [
                  {
                    "code": "CODE-001",
                    "user_id": "17cf3f5f-03bb-492a-bbc4-34a6ad4a4353"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Gift code assigned to user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiftCodeRead"
                },
                "examples": {
                  "assigned": {
                    "value": {
                      "id": "acecf050-e315-4db0-99ca-fdc7af77ec19",
                      "code": "CODE-001",
                      "provider": "amazon",
                      "assigned_to_user_id": "17cf3f5f-03bb-492a-bbc4-34a6ad4a4353",
                      "redeemed": false,
                      "created_at": "2026-10-05T14:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "GiftCodeRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "assigned_to_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "redeemed": {
            "type": "boolean",
            "title": "Redeemed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "provider",
          "assigned_to_user_id",
          "redeemed",
          "created_at"
        ],
        "title": "GiftCodeRead"
      },
      "GiftCodeRedeemRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Code"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "user_id"
        ],
        "title": "GiftCodeRedeemRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}